chore : album extension 설정
