+++ iOS/iOS/Base.lproj/Main.storyboard
@@ -547,7 +547,38 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="XJG-fb-FxN" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="4585" y="-9"/>
+            <point key="canvasLocation" x="4601" y="-9"/>
+        </scene>
+        <!--Album Search View Controller-->
+        <scene sceneID="eCL-tv-P52">
+            <objects>
+                <viewController id="Xp3-Mf-Gjb" customClass="AlbumSearchViewController" customModule="iOS" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="Wny-9d-HTE">
+                        <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="FAv-0B-9aN">
+                                <rect key="frame" x="0.0" y="59" width="393" height="759"/>
+                                <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                                <connections>
+                                    <outlet property="dataSource" destination="Xp3-Mf-Gjb" id="1Ov-aC-YHN"/>
+                                    <outlet property="delegate" destination="Xp3-Mf-Gjb" id="OwV-zf-gw5"/>
+                                </connections>
+                            </tableView>
+                        </subviews>
+                        <viewLayoutGuide key="safeArea" id="pSC-y4-Ksc"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                        <constraints>
+                            <constraint firstItem="FAv-0B-9aN" firstAttribute="top" secondItem="pSC-y4-Ksc" secondAttribute="top" id="cZ5-z0-Ppb"/>
+                            <constraint firstItem="pSC-y4-Ksc" firstAttribute="bottom" secondItem="FAv-0B-9aN" secondAttribute="bottom" id="htd-PF-xe3"/>
+                            <constraint firstItem="FAv-0B-9aN" firstAttribute="leading" secondItem="pSC-y4-Ksc" secondAttribute="leading" id="iN7-3D-2TJ"/>
+                            <constraint firstItem="pSC-y4-Ksc" firstAttribute="trailing" secondItem="FAv-0B-9aN" secondAttribute="trailing" id="zRD-fR-5Os"/>
+                        </constraints>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="kqH-hB-CDD" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="5343" y="-9"/>
         </scene>
         <!--추천-->
         <scene sceneID="5k9-aw-Hl2">
